backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
openai==1.35.3
anthropic==0.28.0
python-dotenv==1.0.1
asyncpg==0.29.0
greenlet==3.0.3
python-multipart==0.0.9
aiohttp==3.9.5
tenacity==8.3.0
structlog==24.2.0
google-generativeai==0.7.2
passlib[bcrypt]==1.7.4
PyJWT==2.8.0